#3 Avatar prop에 따라 수정할수있도록 업데이트
