+++ components/Avatar.js
@@ -1,0 +1,51 @@
+import React from "react";
+import { makeStyles } from "@material-ui/core/styles";
+import Avatar from "@material-ui/core/Avatar";
+
+const useStyles = makeStyles((theme) => ({
+  root: {
+    display: "flex",
+    "& > *": {
+      margin: theme.spacing(1),
+    },
+  },
+  small: {
+    width: theme.spacing(3),
+    height: theme.spacing(3),
+  },
+  medium: {
+    width: theme.spacing(7),
+    height: theme.spacing(7),
+  },
+  large: {
+    width: theme.spacing(13),
+    height: theme.spacing(13),
+  },
+}));
+
+export default function ImageAvatars({ size }) {
+  const classes = useStyles();
+  return (
+    <div className={classes.root}>
+      {size === 0 ? (
+        <Avatar
+          alt="Remy Sharp"
+          src="/static/images/avatar/1.jpg"
+          className={classes.small}
+        />
+      ) : size === 1 ? (
+        <Avatar
+          alt="Remy Sharp"
+          src="/static/images/avatar/1.jpg"
+          className={classes.medium}
+        />
+      ) : (
+        <Avatar
+          alt="Remy Sharp"
+          src="/static/images/avatar/1.jpg"
+          className={classes.large}
+        />
+      )}
+    </div>
+  );
+}
